14496 Move PointOfSaleItemsServiceTests
