+++ WooCommerce/WooCommerceTests/POS/Models/PointOfSaleAggregateModelTests.swift
@@ -61,295 +61,6 @@ struct PointOfSaleAggregateModelTests {
 
     }
 
-    struct ItemListTests {
-        private let itemProvider: MockPOSItemProvider
-        private let sut: PointOfSaleAggregateModel
-
-        init() {
-            itemProvider = MockPOSItemProvider()
-            let itemsService = PointOfSaleItemsService(itemProvider: itemProvider)
-            sut = PointOfSaleAggregateModel(itemsService: itemsService,
-                                            cardPresentPaymentService: MockCardPresentPaymentService(),
-                                            orderService: MockPOSOrderService())
-        }
-
-        @Test func loadInitialItems_requests_first_page() async throws {
-            // Given
-            try #require(sut.itemListState == .initialLoading)
-
-            // When
-            await sut.loadInitialItems()
-
-            // Then
-            #expect(itemProvider.spyLastRequestedPageNumber == 1)
-        }
-
-        @Test func loadInitialItems_results_in_loaded_state() async throws {
-            // Given
-            let expectedItems = MockPOSItemProvider.makeInitialItems()
-            try #require(sut.itemListState == .initialLoading)
-
-            // When
-            await sut.loadInitialItems()
-
-            // Then
-            #expect(sut.itemListState == .loaded(expectedItems))
-        }
-
-        @Test func loadInitialItems_when_called_multiple_times_then_items_are_not_duplicated() async throws {
-            // Given
-            try #require(sut.itemListState == .initialLoading)
-            let expectedItems = MockPOSItemProvider.makeInitialItems()
-
-            // When
-            await sut.loadInitialItems()
-            await sut.loadInitialItems()
-            await sut.loadInitialItems()
-
-            // Then
-            guard case .loaded(let items) = sut.itemListState else {
-                Issue.record("Expected loaded ItemList state, but got \(sut.itemListState)")
-                return
-            }
-            #expect(items.count == expectedItems.count)
-        }
-
-        @Test func reload_results_in_loaded_state() async throws {
-            // Given
-            try #require(sut.itemListState == .initialLoading)
-            let expectedItems = MockPOSItemProvider.makeInitialItems()
-
-            // When
-            await sut.reload()
-
-            // Then
-            guard case .loaded(let items) = sut.itemListState else {
-                Issue.record("Expected loaded ItemList state, but got \(sut.itemListState)")
-                return
-            }
-            #expect(items.count == expectedItems.count)
-        }
-
-        @Test func reload_when_called_multiple_times_then_items_are_not_duplicated() async throws {
-            // Given
-            try #require(sut.itemListState == .initialLoading)
-            let expectedItems = MockPOSItemProvider.makeInitialItems()
-
-            // When
-            await sut.reload()
-            await sut.reload()
-            await sut.reload()
-
-            // Then
-            guard case .loaded(let items) = sut.itemListState else {
-                Issue.record("Expected loaded ItemList state, but got \(sut.itemListState)")
-                return
-            }
-            #expect(items.count == expectedItems.count)
-        }
-
-        @Test func state_starts_as_initialLoading() {
-            // Given/When/Then
-            #expect(sut.itemListState == .initialLoading)
-        }
-
-        @Test func loadItems_when_initial_items_empty_then_state_is_empty() async throws {
-            // Given
-            let itemProvider = MockPOSItemProvider()
-            itemProvider.shouldReturnZeroItems = true
-
-            let itemsService = PointOfSaleItemsService(itemProvider: itemProvider)
-            let sut = PointOfSaleAggregateModel(itemsService: itemsService,
-                                                cardPresentPaymentService: MockCardPresentPaymentService(),
-                                                orderService: MockPOSOrderService())
-
-            try #require(sut.itemListState == .initialLoading)
-
-            // When
-            await sut.loadNextItems()
-
-            // Then
-            #expect(sut.itemListState == .empty)
-        }
-
-        @Test func loadItems_when_initial_items_has_items_then_state_is_loaded_with_initial_items() async throws {
-            // Given
-            let initialItems = MockPOSItemProvider.makeInitialItems()
-            itemProvider.items = initialItems
-
-            try #require(sut.itemListState == .initialLoading)
-
-            // When
-            await sut.loadNextItems()
-
-            // Then
-            #expect(sut.itemListState == .loaded(initialItems))
-        }
-
-        @Test func loadItems_when_simulateFetchNextPage_then_state_is_loaded_with_expected_items() async throws {
-            // Given
-            let initialItems = MockPOSItemProvider.makeInitialItems()
-            itemProvider.items = initialItems
-            itemProvider.shouldSimulateTwoPages = true
-
-            // When
-            await sut.loadNextItems()
-
-            // Then
-            guard case .loaded(let items) = sut.itemListState else {
-                Issue.record("Expected loaded ItemList state, but got \(sut.itemListState)")
-                return
-            }
-            #expect(items.count == 4)
-        }
-
-        @Test func loadNextItems_requests_second_page() async throws {
-            // Given
-            try #require(sut.itemListState == .initialLoading)
-
-            // When
-            await sut.loadNextItems()
-
-            // Then
-            #expect(itemProvider.spyLastRequestedPageNumber == 2)
-        }
-
-        @Test func loadInitialItems_when_no_items_then_state_is_loaded_empty() async throws {
-            // Given
-            let itemProvider = MockPOSItemProvider()
-            itemProvider.shouldReturnZeroItems = true
-            let itemsService = PointOfSaleItemsService(itemProvider: itemProvider)
-            let sut = PointOfSaleAggregateModel(itemsService: itemsService,
-                                                cardPresentPaymentService: MockCardPresentPaymentService(),
-                                                orderService: MockPOSOrderService())
-
-            try #require(sut.itemListState == .initialLoading)
-
-            // When
-            await sut.loadInitialItems()
-
-            // Then
-            #expect(sut.itemListState == .empty)
-        }
-
-        @Test func loadInitialItems_when_itemProvider_throws_error_then_state_is_error() async throws {
-            // Given
-            itemProvider.shouldThrowError = true
-            let expectedError = PointOfSaleErrorState(title: "Error loading products",
-                                                      subtitle: "Give it another go?",
-                                                      buttonText: "Retry")
-            try #require(sut.itemListState == .initialLoading)
-
-            // When
-            await sut.loadInitialItems()
-
-            // Then
-            #expect(sut.itemListState == .error(expectedError))
-        }
-
-        @Test func loadNextItems_when_itemProvider_throws_error_then_state_is_error() async throws {
-            // Given
-            itemProvider.shouldThrowError = true
-            let expectedError = PointOfSaleErrorState(title: "Error loading products",
-                                                      subtitle: "Give it another go?",
-                                                      buttonText: "Retry")
-            try #require(sut.itemListState == .initialLoading)
-
-            // When
-            await sut.loadNextItems()
-
-            // Then
-            #expect(sut.itemListState == .error(expectedError))
-        }
-
-        @Test func loadNextItems_after_itemProvider_throws_error_then_the_same_page_is_requested_next() async throws {
-            // Given
-            itemProvider.shouldThrowError = true
-            await sut.loadNextItems()
-            try #require(itemProvider.spyLastRequestedPageNumber == 2)
-            itemProvider.spyLastRequestedPageNumber = 0
-
-            // When
-            await sut.loadNextItems()
-
-            // Then
-            #expect(itemProvider.spyLastRequestedPageNumber == 2)
-        }
-
-        @Test func reload_results_in_state_loaded_with_expected_items() async throws {
-            // Given
-            try #require(sut.itemListState == .initialLoading)
-            let expectedItems = MockPOSItemProvider.makeInitialItems()
-
-            // When
-            await sut.reload()
-
-            // Then
-            #expect(sut.itemListState == .loaded(expectedItems))
-        }
-
-        @Test func reload_requests_first_page() async throws {
-            // Given
-            await sut.loadNextItems()
-            try #require(itemProvider.spyLastRequestedPageNumber == 2)
-
-            // When
-            await sut.reload()
-
-            // Then
-            #expect(itemProvider.spyLastRequestedPageNumber == 1)
-        }
-
-        @Test func loadNextItems_when_next_page_is_out_of_range_then_receives_error() async throws {
-            // Given
-            await sut.loadInitialItems()
-            try #require(itemProvider.spyLastRequestedPageNumber == 1)
-            let expectedError = PointOfSaleErrorState(title: "Error loading products",
-                                                      subtitle: "Give it another go?",
-                                                      buttonText: "Retry")
-
-            // When
-            itemProvider.simulateNextPageIsOutOfRange()
-            await sut.loadNextItems()
-
-            // Then
-            guard case .error = sut.itemListState else {
-                Issue.record("Expected error state, but got \(sut.itemListState)")
-                return
-            }
-            #expect(sut.itemListState == .error(expectedError))
-        }
-
-        @Test func loadNextItems_when_next_page_is_out_of_range_then_the_same_page_is_requested_next() async throws {
-            // Given
-            await sut.loadInitialItems()
-            try #require(itemProvider.spyLastRequestedPageNumber == 1)
-
-            // When
-            itemProvider.simulateNextPageIsOutOfRange()
-            await sut.loadNextItems()
-
-            // Then
-            try #require(itemProvider.spyLastRequestedPageNumber == 1)
-        }
-
-        @Test func reload_when_itemProvider_throws_error_then_state_is_error() async throws {
-            // Given
-            itemProvider.shouldThrowError = true
-            let expectedError = PointOfSaleErrorState(title: "Error loading products",
-                                                      subtitle: "Give it another go?",
-                                                      buttonText: "Retry")
-
-            try #require(sut.itemListState == .initialLoading)
-
-            // When
-            await sut.reload()
-
-            // Then
-            #expect(sut.itemListState == .error(expectedError))
-        }
-    }
-
     struct CartTests {
         let sut: PointOfSaleAggregateModel
         private let analytics: WooAnalytics!
